Moved ReducedGraphParent building to a new module
